+++ src/main/java/chess/controller/ChessController.java
@@ -1,5 +1,8 @@
 package chess.controller;
 
+import chess.controller.command.Command;
+import chess.controller.command.CommandType;
+import chess.controller.command.StartCommand;
 import chess.view.InputView;
 import chess.view.OutputView;
