+++ src/main/java/chess/controller/command/EndCommand.java
@@ -1,4 +1,4 @@
-package chess.controller;
+package chess.controller.command;
 
 import chess.domain.ChessGame;
 
@@ -6,12 +6,14 @@ import java.util.List;
 
 public class EndCommand extends Command{
 
+    private static final String INVALID_EXECUTE_MESSAGE = "게임이 종료되었습니다.";
+
     protected EndCommand(ChessGame chessGame) {
         super(chessGame, CommandType.END);
     }
 
     @Override
     public Command execute(List<String> input) {
-        throw new IllegalArgumentException("게임이 종료되었습니다.");
+        throw new IllegalArgumentException(INVALID_EXECUTE_MESSAGE);
     }
 }
