+++ src/main/java/chess/view/OutputView.java
@@ -6,17 +6,19 @@ import chess.domain.board.Position;
 import chess.domain.board.RankCoordinate;
 import chess.domain.piece.Piece;
 
+import java.util.List;
 import java.util.Map;
 
 public class OutputView {
 
     private static final String ERROR_START_MESSAGE = "[ERROR]: ";
+    private static final String LINE_SEPARATOR = System.lineSeparator();
 
     public void printStart() {
-        System.out.println("> 체스 게임을 시작합니다.\n"
-                + "> 게임 시작 : start\n"
-                + "> 게임 종료 : end\n"
-                + "> 게임 이동 : move source위치 target위치 - 예. move b2 b3");
+        String startMessage = "> 체스 게임을 시작합니다."  + LINE_SEPARATOR + "> 게임 시작 : start" + LINE_SEPARATOR
+                + "> 게임 종료 : end" + LINE_SEPARATOR + "> 게임 이동 : move source위치 target위치 - 예. move b2 b3";
+
+        System.out.println(startMessage);
     }
 
     public void printBoard(Map<Position, Piece> boards) {
