+++ src/test/java/chess/controller/command/EndCommandTest.java
@@ -1,5 +1,8 @@
-package chess.controller;
+package chess.controller.command;
 
+import chess.controller.command.Command;
+import chess.controller.command.CommandType;
+import chess.controller.command.EndCommand;
 import chess.domain.ChessGame;
 import chess.domain.board.BoardFactory;
 import chess.domain.piece.Pawn;
